refactor: 이상형 월드컵 폴더 구조
