+++ src/screens/worldcup/card/CardWorldCupScreen.styled.ts
@@ -37,7 +37,7 @@ export const CardContainer = styled.View`
   box-shadow: 0px 4px 4px rgba(77, 77, 77, 0.25);
   border-radius: 15px;
   height: 45%;
-  width: 100%
+  width: 100%;
 `;
 
 export const CardImgContainer = styled.TouchableOpacity`
