+++ src/screens/worldcup/information/WorldCupInfoScreen.tsx
@@ -9,9 +9,9 @@ import {
   WorldCupInfoImgContainer,
   WorldCupTitle,
 } from './WorldCupInfoScreen.styled';
-import { WorldCupIcon } from '../../assets';
+import { WorldCupIcon } from '../../../assets';
 
-export default function WorldcupInfoScreen({ navigation }: any) {
+export default function WorldCupInfoScreen({ navigation }: any) {
   return (
     <Container>
       <ScrollViewContainer
@@ -22,12 +22,12 @@ export default function WorldcupInfoScreen({ navigation }: any) {
         }}
       >
         <WorldCupInfoImgContainer>
-          <WorldCupIcon width="100%" height="100%"/>
+          <WorldCupIcon width="100%" height="100%" />
         </WorldCupInfoImgContainer>
         <WorldCupTitle>이상형 월드컵</WorldCupTitle>
         <WorldCupExplaination>내게 맞는 카드</WorldCupExplaination>
         <WorldCupExplaination>일대일 비교로 간편하게</WorldCupExplaination>
-        <EnterButton onPress={() => navigation.navigate("WorldCup")}>
+        <EnterButton onPress={() => navigation.navigate('WorldCup')}>
           <EnterText>지금 시작!</EnterText>
         </EnterButton>
         <TouchableOpacity>
