+++ src/screens/worldcup/select/WorldCupSelectScreen.styled.ts
