+++ src/screens/worldcup/winner/CardWorldCupWinnerScreen.styled.ts
